Move serializer caching from adapter
