+++ lib/active_model_serializers/adapter/base.rb
@@ -23,8 +23,8 @@ module ActiveModelSerializers
         hash
       end
 
-      def fragment_cache(*_args)
-        fail NotImplementedError, 'This is an abstract method. Should be implemented at the concrete adapter.'
+      def fragment_cache(cached_hash, non_cached_hash)
+        non_cached_hash.merge cached_hash
       end
 
       def cache_check(serializer)
